requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
orjson==3.10.7
ujson==5.10.0
typing-extensions==4.12.2
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.0.8
PyJWT==2.9.0
openpyxl==3.1.5
xlrd==2.0.1
numpy==2.1.2
pandas==2.2.3
supabase==2.10.0
passlib[bcrypt]==1.7.4
chromadb==0.5.11
sentence-transformers==3.2.1
google-generativeai==0.8.3

